feat: refactor folders
